+++ src/main/java/ru/yandex/practicum/stellarburger/pages/RegisterPage.java
@@ -7,7 +7,7 @@ import io.qameta.allure.Step;
 import org.openqa.selenium.support.FindBy;
 import org.openqa.selenium.support.How;
 
-public class RegisterPage {
+public class RegisterPage extends HeaderPage {
     public static final String INCORRECT_PASSWORD_MESSAGE = "Некорректный пароль";
 
     @FindBy(how = How.XPATH, using = ".//label[text()='Имя']/following-sibling::input")
